+++ apps/backend/src/app/api/latest/auth/oauth/callback/[provider_id]/route.tsx
@@ -1,14 +1,13 @@
 import { usersCrudHandlers } from "@/app/api/latest/users/crud";
 import { getAuthContactChannel } from "@/lib/contact-channel";
-import { getProject } from "@/lib/projects";
 import { validateRedirectUrl } from "@/lib/redirect-urls";
+import { Tenancy, getTenancy } from "@/lib/tenancies";
 import { oauthCookieSchema } from "@/lib/tokens";
 import { getProvider, oauthServer } from "@/oauth";
 import { prismaClient } from "@/prisma-client";
 import { createSmartRouteHandler } from "@/route-handlers/smart-route-handler";
 import { InvalidClientError, InvalidScopeError, Request as OAuthRequest, Response as OAuthResponse } from "@node-oauth/oauth2-server";
 import { KnownError, KnownErrors } from "@stackframe/stack-shared";
-import { ProjectsCrud } from "@stackframe/stack-shared/dist/interface/crud/projects";
 import { yupMixed, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
 import { StackAssertionError, StatusError, captureError } from "@stackframe/stack-shared/dist/utils/errors";
 import { deindent, extractScopes } from "@stackframe/stack-shared/dist/utils/strings";
@@ -16,8 +15,8 @@ import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 import { oauthResponseToSmartResponse } from "../../oauth-helpers";
 
-const redirectOrThrowError = (error: KnownError, project: ProjectsCrud["Admin"]["Read"], errorRedirectUrl?: string) => {
-  if (!errorRedirectUrl || !validateRedirectUrl(errorRedirectUrl, project.config.domains, project.config.allow_localhost)) {
+const redirectOrThrowError = (error: KnownError, tenancy: Tenancy, errorRedirectUrl?: string) => {
+  if (!errorRedirectUrl || !validateRedirectUrl(errorRedirectUrl, tenancy.config.domains, tenancy.config.allow_localhost)) {
     throw error;
   }
 
@@ -72,7 +71,7 @@ const handler = createSmartRouteHandler({
     }
 
     const {
-      projectId,
+      tenancyId,
       innerCodeVerifier,
       type,
       projectUserId,
@@ -81,9 +80,9 @@ const handler = createSmartRouteHandler({
       afterCallbackRedirectUrl,
     } = outerInfo;
 
-    const project = await getProject(projectId);
-    if (!project) {
-      throw new StackAssertionError("Project in outerInfo not found; has it been deleted?", { projectId });
+    const tenancy = await getTenancy(tenancyId);
+    if (!tenancy) {
+      throw new StackAssertionError("Tenancy in outerInfo not found; has it been deleted?", { tenancyId });
     }
 
     try {
@@ -91,7 +90,7 @@ const handler = createSmartRouteHandler({
         throw new KnownErrors.OuterOAuthTimeout();
       }
 
-      const provider = project.config.oauth_providers.find((p) => p.id === params.provider_id);
+      const provider = tenancy.config.oauth_providers.find((p) => p.id === params.provider_id);
       if (!provider || !provider.enabled) {
         throw new KnownErrors.OAuthProviderNotFoundOrNotEnabled();
       }
@@ -109,7 +108,7 @@ const handler = createSmartRouteHandler({
         });
       } catch (error) {
         if (error instanceof KnownErrors['OAuthProviderAccessDenied']) {
-          redirectOrThrowError(error, project, errorRedirectUrl);
+          redirectOrThrowError(error, tenancy, errorRedirectUrl);
         }
         throw error;
       }
@@ -123,8 +122,8 @@ const handler = createSmartRouteHandler({
 
         const user = await prismaClient.projectUser.findUnique({
           where: {
-            projectId_projectUserId: {
-              projectId,
+            tenancyId_projectUserId: {
+              tenancyId,
               projectUserId,
             },
           },
@@ -151,7 +150,7 @@ const handler = createSmartRouteHandler({
         body: {},
         method: "GET",
         query: {
-          client_id: outerInfo.projectId,
+          client_id: tenancy.project.id,
           client_secret: outerInfo.publishableClientKey,
           redirect_uri: outerInfo.redirectUri,
           state: outerInfo.state,
@@ -167,7 +166,7 @@ const handler = createSmartRouteHandler({
         if (tokenSet.refreshToken) {
           await prismaClient.oAuthToken.create({
             data: {
-              projectId: outerInfo.projectId,
+              tenancyId: outerInfo.tenancyId,
               oAuthProviderConfigId: provider.id,
               refreshToken: tokenSet.refreshToken,
               providerAccountId: userInfo.accountId,
@@ -178,7 +177,7 @@ const handler = createSmartRouteHandler({
 
         await prismaClient.oAuthAccessToken.create({
           data: {
-            projectId: outerInfo.projectId,
+            tenancyId: outerInfo.tenancyId,
             oAuthProviderConfigId: provider.id,
             accessToken: tokenSet.accessToken,
             providerAccountId: userInfo.accountId,
@@ -198,8 +197,8 @@ const handler = createSmartRouteHandler({
               handle: async () => {
                 const oldAccount = await prismaClient.projectUserOAuthAccount.findUnique({
                   where: {
-                    projectId_oauthProviderConfigId_providerAccountId: {
-                      projectId: outerInfo.projectId,
+                    tenancyId_oauthProviderConfigId_providerAccountId: {
+                      tenancyId: outerInfo.tenancyId,
                       oauthProviderConfigId: provider.id,
                       providerAccountId: userInfo.accountId,
                     },
@@ -227,15 +226,15 @@ const handler = createSmartRouteHandler({
                         providerConfig: {
                           connect: {
                             projectConfigId_id: {
-                              projectConfigId: project.config.id,
+                              projectConfigId: tenancy.config.id,
                               id: provider.id,
                             },
                           },
                         },
                         projectUser: {
                           connect: {
-                            projectId_projectUserId: {
-                              projectId: outerInfo.projectId,
+                            tenancyId_projectUserId: {
+                              tenancyId: outerInfo.tenancyId,
                               projectUserId: projectUserId,
                             },
                           },
@@ -266,7 +265,7 @@ const handler = createSmartRouteHandler({
 
                   // ========================== sign up user ==========================
 
-                  if (!project.config.sign_up_enabled) {
+                  if (!tenancy.config.sign_up_enabled) {
                     throw new KnownErrors.SignUpNotEnabled();
                   }
 
@@ -277,7 +276,7 @@ const handler = createSmartRouteHandler({
                     const oldContactChannel = await getAuthContactChannel(
                       prismaClient,
                       {
-                        projectId: outerInfo.projectId,
+                        tenancyId: outerInfo.tenancyId,
                         type: 'EMAIL',
                         value: userInfo.email,
                       }
@@ -291,7 +290,7 @@ const handler = createSmartRouteHandler({
                   }
 
                   const newAccount = await usersCrudHandlers.adminCreate({
-                    project,
+                    tenancy,
                     data: {
                       display_name: userInfo.displayName,
                       profile_image_url: userInfo.profileImageUrl || undefined,
@@ -339,7 +338,7 @@ const handler = createSmartRouteHandler({
       return oauthResponseToSmartResponse(oauthResponse);
     } catch (error) {
       if (error instanceof KnownError) {
-        redirectOrThrowError(error, project, errorRedirectUrl);
+        redirectOrThrowError(error, tenancy, errorRedirectUrl);
       }
       throw error;
     }