+++ apps/backend/src/app/api/latest/teams/crud.tsx
@@ -43,7 +43,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
       throw new KnownErrors.UserAuthenticationRequired;
     }
 
-    if (auth.type === 'client' && !auth.project.config.client_team_creation_enabled) {
+    if (auth.type === 'client' && !auth.tenancy.config.client_team_creation_enabled) {
       throw new StatusError(StatusError.Forbidden, 'Client team creation is disabled for this project');
     }
 
@@ -55,7 +55,9 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
       const db = await tx.team.create({
         data: {
           displayName: data.display_name,
-          projectId: auth.project.id,
+          mirroredProjectId: auth.project.id,
+          mirroredBranchId: auth.tenancy.branchId,
+          tenancyId: auth.tenancy.id,
           profileImageUrl: data.profile_image_url,
           clientMetadata: data.client_metadata === null ? Prisma.JsonNull : data.client_metadata,
           clientReadOnlyMetadata: data.client_read_only_metadata === null ? Prisma.JsonNull : data.client_read_only_metadata,
@@ -81,7 +83,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
 
       if (addUserId) {
         await addUserToTeam(tx, {
-          project: auth.project,
+          tenancy: auth.tenancy,
           teamId: db.teamId,
           userId: addUserId,
           type: 'creator',
@@ -103,7 +105,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
   onRead: async ({ params, auth }) => {
     if (auth.type === 'client') {
       await ensureTeamMembershipExists(prismaClient, {
-        projectId: auth.project.id,
+        tenancyId: auth.tenancy.id,
         teamId: params.team_id,
         userId: auth.user?.id ?? throwErr(new KnownErrors.UserAuthenticationRequired),
       });
@@ -111,8 +113,8 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
 
     const db = await prismaClient.team.findUnique({
       where: {
-        projectId_teamId: {
-          projectId: auth.project.id,
+        tenancyId_teamId: {
+          tenancyId: auth.tenancy.id,
           teamId: params.team_id,
         },
       },
@@ -132,7 +134,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
 
       if (auth.type === 'client') {
         await ensureUserTeamPermissionExists(tx, {
-          project: auth.project,
+          tenancy: auth.tenancy,
           teamId: params.team_id,
           userId: auth.user?.id ?? throwErr(new KnownErrors.UserAuthenticationRequired),
           permissionId: "$update_team",
@@ -141,12 +143,12 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
         });
       }
 
-      await ensureTeamExists(tx, { projectId: auth.project.id, teamId: params.team_id });
+      await ensureTeamExists(tx, { tenancyId: auth.tenancy.id, teamId: params.team_id });
 
       return await tx.team.update({
         where: {
-          projectId_teamId: {
-            projectId: auth.project.id,
+          tenancyId_teamId: {
+            tenancyId: auth.tenancy.id,
             teamId: params.team_id,
           },
         },
@@ -173,7 +175,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
     await retryTransaction(async (tx) => {
       if (auth.type === 'client') {
         await ensureUserTeamPermissionExists(tx, {
-          project: auth.project,
+          tenancy: auth.tenancy,
           teamId: params.team_id,
           userId: auth.user?.id ?? throwErr(new KnownErrors.UserAuthenticationRequired),
           permissionId: "$delete_team",
@@ -181,12 +183,12 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
           recursive: true,
         });
       }
-      await ensureTeamExists(tx, { projectId: auth.project.id, teamId: params.team_id });
+      await ensureTeamExists(tx, { tenancyId: auth.tenancy.id, teamId: params.team_id });
 
       await tx.team.delete({
         where: {
-          projectId_teamId: {
-            projectId: auth.project.id,
+          tenancyId_teamId: {
+            tenancyId: auth.tenancy.id,
             teamId: params.team_id,
           },
         },
@@ -211,7 +213,7 @@ export const teamsCrudHandlers = createLazyProxy(() => createCrudHandlers(teamsC
 
     const db = await prismaClient.team.findMany({
       where: {
-        projectId: auth.project.id,
+        tenancyId: auth.tenancy.id,
         ...query.user_id ? {
           teamMembers: {
             some: {