+++ apps/backend/src/lib/emails.tsx
@@ -3,7 +3,6 @@ import { prismaClient } from '@/prisma-client';
 import { traceSpan } from '@/utils/telemetry';
 import { TEditorConfiguration } from '@stackframe/stack-emails/dist/editor/documents/editor/core';
 import { EMAIL_TEMPLATES_METADATA, renderEmailTemplate } from '@stackframe/stack-emails/dist/utils';
-import { ProjectsCrud } from '@stackframe/stack-shared/dist/interface/crud/projects';
 import { UsersCrud } from '@stackframe/stack-shared/dist/interface/crud/users';
 import { getEnvVariable } from '@stackframe/stack-shared/dist/utils/env';
 import { StackAssertionError, captureError } from '@stackframe/stack-shared/dist/utils/errors';
@@ -12,6 +11,8 @@ import { runAsynchronously, wait } from '@stackframe/stack-shared/dist/utils/pro
 import { Result } from '@stackframe/stack-shared/dist/utils/results';
 import { typedToUppercase } from '@stackframe/stack-shared/dist/utils/strings';
 import nodemailer from 'nodemailer';
+import { Tenancy } from './tenancies';
+
 export async function getEmailTemplate(projectId: string, type: keyof typeof EMAIL_TEMPLATES_METADATA) {
   const project = await getProject(projectId);
   if (!project) {
@@ -247,24 +248,24 @@ export async function sendEmail(options: SendEmailOptions) {
 }
 
 export async function sendEmailFromTemplate(options: {
-  project: ProjectsCrud["Admin"]["Read"],
+  tenancy: Tenancy,
   user: UsersCrud["Admin"]["Read"] | null,
   email: string,
   templateType: keyof typeof EMAIL_TEMPLATES_METADATA,
   extraVariables: Record<string, string | null>,
   version?: 1 | 2,
 }) {
-  const template = await getEmailTemplateWithDefault(options.project.id, options.templateType, options.version);
+  const template = await getEmailTemplateWithDefault(options.tenancy.project.id, options.templateType, options.version);
 
   const variables = filterUndefined({
-    projectDisplayName: options.project.display_name,
+    projectDisplayName: options.tenancy.project.display_name,
     userDisplayName: options.user?.display_name || undefined,
     ...filterUndefined(options.extraVariables),
   });
   const { subject, html, text } = renderEmailTemplate(template.subject, template.content, variables);
 
   await sendEmail({
-    emailConfig: await getEmailConfig(options.project),
+    emailConfig: await getEmailConfig(options.tenancy),
     to: options.email,
     subject,
     html,
@@ -272,8 +273,8 @@ export async function sendEmailFromTemplate(options: {
   });
 }
 
-async function getEmailConfig(project: ProjectsCrud["Admin"]["Read"]): Promise<EmailConfig> {
-  const projectEmailConfig = project.config.email_config;
+async function getEmailConfig(tenancy: Tenancy): Promise<EmailConfig> {
+  const projectEmailConfig = tenancy.config.email_config;
 
   if (projectEmailConfig.type === 'shared') {
     return {
@@ -282,13 +283,13 @@ async function getEmailConfig(project: ProjectsCrud["Admin"]["Read"]): Promise<E
       username: getEnvVariable('STACK_EMAIL_USERNAME'),
       password: getEnvVariable('STACK_EMAIL_PASSWORD'),
       senderEmail: getEnvVariable('STACK_EMAIL_SENDER'),
-      senderName: project.display_name,
+      senderName: tenancy.project.display_name,
       secure: isSecureEmailPort(getEnvVariable('STACK_EMAIL_PORT')),
       type: 'shared',
     };
   } else {
     if (!projectEmailConfig.host || !projectEmailConfig.port || !projectEmailConfig.username || !projectEmailConfig.password || !projectEmailConfig.sender_email || !projectEmailConfig.sender_name) {
-      throw new StackAssertionError("Email config is not complete despite not being shared. This should never happen?", { projectId: project.id, emailConfig: projectEmailConfig });
+      throw new StackAssertionError("Email config is not complete despite not being shared. This should never happen?", { projectId: tenancy.id, emailConfig: projectEmailConfig });
     }
     return {
       host: projectEmailConfig.host,